add lottie, add upvote

+++ pages/api/index.ts
@@ -28,6 +28,8 @@ export default async function handler(
       break
     case "POST":
       try {
+        if (!session)
+          return res.status(400).json({ error: "not authenticated" })
         const message = await Message.create(req.body)
         res.status(201).json(message)
       } catch (error) {
@@ -36,11 +38,21 @@ export default async function handler(
       break
     case "PUT":
       try {
-        const { _id, update } = req.body
-        const message = await Message.updateOne(
-          { _id: _id },
-          { $set: { content: update } }
-        )
+        if (!session)
+          return res.status(400).json({ error: "not authenticated" })
+        const { _id, update, upvote } = req.body
+        let message
+        if (upvote) {
+          message = await Message.updateOne(
+            { _id: _id },
+            { $inc: { upvote: 1 } }
+          )
+        } else
+          message = await Message.updateOne(
+            { _id: _id },
+            { $set: { content: update } }
+          )
+
         res.status(200).json({ success: message })
       } catch (error) {
         res.status(400).json({ success: false })
@@ -48,6 +60,8 @@ export default async function handler(
       break
     case "DELETE":
       try {
+        if (!session)
+          return res.status(400).json({ error: "not authenticated" })
         const id = req.body._id
         const message = await Message.deleteOne({ _id: id })
         res.status(200).json({ success: message })
